test(utils): fix tests of `getProp` and `setProp`
